* fixed compiler warnings
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@3447 599ab11a-0111-0410-8a50-d7a5b0953570

+++ librapi2/src/rapi/file_management.c
@@ -224,7 +224,7 @@ DWORD _CeGetSpecialFolderPath(
 		LPWSTR lpBuffer)
 {
 	RapiContext* context = rapi_context_current();
-	size_t string_length = nBufferLength;
+	DWORD string_length = nBufferLength;
 
 	rapi_context_begin_command(context, 0x44);
 	rapi_buffer_write_uint32(context->send_buffer, (uint32_t)nFolder);
